+++ src/libs/ascent/runtimes/flow_filters/ascent_runtime_vtkh_filters.hpp
@@ -451,6 +451,19 @@ public:
 };
 
 //-----------------------------------------------------------------------------
+class ASCENT_API VTKHTransform : public ::flow::Filter
+{
+public:
+    VTKHTransform();
+    virtual ~VTKHTransform();
+
+    virtual void   declare_interface(conduit::Node &i);
+    virtual bool   verify_params(const conduit::Node &params,
+                                 conduit::Node &info);
+    virtual void   execute();
+};
+
+//-----------------------------------------------------------------------------
 class ASCENT_API VTKHTriangulate : public ::flow::Filter
 {
 public: