+++ src/libs/vtkh/filters/PointTransform.cpp
@@ -5,61 +5,114 @@
 namespace vtkh
 {
 
+//---------------------------------------------------------------------------//
 PointTransform::PointTransform()
 {
   ResetTransform();
 }
 
+//---------------------------------------------------------------------------//
 PointTransform::~PointTransform()
 {
 
 }
 
-void PointTransform::ResetTransform()
+//---------------------------------------------------------------------------//
+void
+PointTransform::ResetTransform()
 {
   vtkm::MatrixIdentity(m_transform);
 }
 
+//---------------------------------------------------------------------------//
 void
 PointTransform::SetTranslation(const double& tx,
                                const double& ty,
                                const double& tz)
 {
-  vtkm::Matrix<double,4,4> matrix  = vtkm::Transform3DTranslate(tx, ty, tz);
+  vtkm::Matrix<double,4,4> matrix = vtkm::Transform3DTranslate(tx, ty, tz);
   m_transform = vtkm::MatrixMultiply(m_transform, matrix);
 }
 
-void PointTransform::SetRotation(const double& angleDegrees,
-                                 const vtkm::Vec<double, 3>& axis)
+//---------------------------------------------------------------------------//
+void
+PointTransform::SetRotation(const double& angleDegrees,
+                            const double& axisX,
+                            const double& axisY,
+                            const double& axisZ)
 {
-  vtkm::Matrix<double,4,4> matrix = vtkm::Transform3DRotate(angleDegrees, axis);
+  vtkm::Matrix<double,4,4> matrix = vtkm::Transform3DRotate(angleDegrees,
+                                                            axisX,
+                                                            axisY,
+                                                            axisZ);
   m_transform = vtkm::MatrixMultiply(m_transform, matrix);
 }
 
-void PointTransform::SetTransform(const vtkm::Matrix<double, 4, 4>& mtx)
+//---------------------------------------------------------------------------//
+void
+PointTransform::SetTransform(const double *matrix_values)
+{
+  // Note: row vs col vs matrix vs array, what is the best
+  //       order for users to provide flat values to matrix?
+  //       This order is decent, was able to throw values in
+  //       from example matrices in the wild easily.
+
+  m_transform[0][0] = matrix_values[0];
+  m_transform[0][1] = matrix_values[1];
+  m_transform[0][2] = matrix_values[2];
+  m_transform[0][3] = matrix_values[3];
+
+  m_transform[1][0] = matrix_values[4];
+  m_transform[1][1] = matrix_values[5];
+  m_transform[1][2] = matrix_values[6];
+  m_transform[1][3] = matrix_values[7];
+
+  m_transform[2][0] = matrix_values[8];
+  m_transform[2][1] = matrix_values[9];
+  m_transform[2][2] = matrix_values[10];
+  m_transform[2][3] = matrix_values[11];
+
+  m_transform[3][0] = matrix_values[12];
+  m_transform[3][1] = matrix_values[13];
+  m_transform[3][2] = matrix_values[14];
+  m_transform[3][3] = matrix_values[15];
+}
+
+
+//---------------------------------------------------------------------------//
+void
+PointTransform::SetTransform(const vtkm::Matrix<double, 4, 4>& mtx)
 {
   m_transform = mtx;
 }
 
-void PointTransform::SetScale(const double& sx,
-                              const double& sy,
-                              const double& sz)
+//---------------------------------------------------------------------------//
+void
+PointTransform::SetScale(const double& sx,
+                         const double& sy,
+                         const double& sz)
 {
   vtkm::Matrix<double,4,4> matrix = vtkm::Transform3DScale(sx, sy, sz);
   m_transform = vtkm::MatrixMultiply(m_transform, matrix);
 }
 
-void PointTransform::PreExecute()
+//---------------------------------------------------------------------------//
+void
+PointTransform::PreExecute()
 {
   Filter::PreExecute();
 }
 
-void PointTransform::PostExecute()
+//---------------------------------------------------------------------------//
+void
+PointTransform::PostExecute()
 {
   Filter::PostExecute();
 }
 
-void PointTransform::DoExecute()
+//---------------------------------------------------------------------------//
+void
+PointTransform::DoExecute()
 {
   this->m_output = new DataSet();
   const int num_domains = this->m_input->GetNumberOfDomains();
@@ -78,6 +131,7 @@ void PointTransform::DoExecute()
   }
 }
 
+//---------------------------------------------------------------------------//
 std::string
 PointTransform::GetName() const
 {